Fixed issue in Python-3 for test_plot_quadrature

+++ openmoc/plotter.py
@@ -1238,7 +1238,7 @@ def plot_quadrature(solver, get_figure=False):
     quad = track_generator.getQuadrature()
     num_azim = track_generator.getNumAzim()
     azim_spacing = track_generator.getDesiredAzimSpacing()
-    num_polar_2 = quad.getNumPolarAngles() / 2
+    num_polar_2 = int(quad.getNumPolarAngles() / 2)
     phis = np.zeros(num_azim/4)
     thetas = np.zeros(num_polar_2)
